+++ ConstraintLens/lib/messaging.py
@@ -19,6 +19,7 @@ ACTION_OPEN_EDIT_DIALOG = "openEditDialog"
 ACTION_EDIT_PARAMETER = "editParameter"
 ACTION_SET_AUTO_ZOOM = "setAutoZoom"
 ACTION_SET_PALETTE_HEIGHT = "setPaletteHeight"
+ACTION_SET_AUTO_HIDE = "setAutoHide"
 
 PY_ACTION_DATA = "data"
 PY_ACTION_NO_ACTIVE_SKETCH = "noActiveSketch"